tmp/ref.sub.html

<!DOCTYPE html>
<html>
	<head>
		<meta charset="utf-8">
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script src="./help.sub.js"></script>
	</head>
	<body>
<script>
'use strict';
const createObjectUrl = (t, content, mimeType) => {
	const encoder = new TextEncoder();
	const url = URL.createObjectURL(new Blob([encoder.encode(content)], {type: mimeType}));

	t.add_cleanup(() => URL.revokeObjectURL(url));

	return url;
};

const URL_COMPLETE = new URL('/tmp/child.html', location).href;
const URL_DOMAIN = location.origin + '/';

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	createMeta(parent, 'referrer', 'origin');
	parent.document.body.appendChild(child);

	assert_equals(await readReferer(child.contentWindow), URL_DOMAIN);
}, 'src-less iframe: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	parent.document.body.appendChild(child);
	createMeta(parent, 'referrer', 'origin');

	assert_equals(await readReferer(child.contentWindow), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'src-less iframe: child independence from parent');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	parent.document.body.appendChild(child);
	createMeta(child.contentWindow, 'referrer', 'origin');

	assert_equals(await readReferer(child.contentWindow), URL_DOMAIN);
	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'src-less iframe: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	createMeta(parent, 'referrer', 'origin');
	parent.document.body.appendChild(child);
	createMeta(child.contentWindow, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(child.contentWindow), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'src-less iframe: parent independence from child (child de-escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	createMeta(parent, 'referrer', 'origin');

	const child = await createIframe(t, parent);

	assert_equals(await readReferer(child), URL_DOMAIN);
}, 'popup: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = await createPopup(t, parent);

	createMeta(parent, 'referrer', 'origin');

	assert_equals(await readReferer(child), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'popup: child independence from parent');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = await createPopup(t, parent);

	createMeta(child, 'referrer', 'origin');

	assert_equals(await readReferer(child), URL_DOMAIN);
	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'popup: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	createMeta(parent, 'referrer', 'origin');
	const child = await createPopup(t, parent);

	createMeta(child, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(child), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'popup: parent independence from child (child de-escalates)');

async function send(win, message) {
	return new Promise((resolve) => {
		addEventListener('message', (event) => {
			resolve(event.data);
		});
		win.postMessage(message, '*');
	});
}

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await dataNav(parent);

	assert_equals(await send(parent, 'readReferer'), URL_DOMAIN);
}, 'data: URL: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);

	await dataNav(parent);
	await send(parent, ['createMeta', 'referrer', 'origin']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'data: URL: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await dataNav(parent);
	await send(parent, ['createMeta', 'referrer', 'unsafe-url']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'data: URL: parent independence from child (child de-escalates)');
</script>
	</body>
</html>
